+++ src/views/home/components/clock-info.vue
@@ -41,75 +41,11 @@
 </template>
 
 <script setup lang="ts">
-import { checkoutInfo, checkoutList, getUserInfoService } from '@/services';
 import { clockInfoDataList, weatherInfoList } from '../configs'
-import { User } from '../type/index'
 import { userMainStore } from '../store'
-import { Message } from '@arco-design/web-vue';
 
 const userMain = userMainStore()
-const { userList,selfUser,dataList } = storeToRefs(userMain)
-
-const id = localStorage.getItem('id') || ""
-const token = localStorage.getItem("token") || ""
-//更新个人信息
-onMounted(async()=>{
-    try{
-        const res = await getUserInfoService({id,token})
-        const info = res.response?.data
-        if(info){
-            selfUser.value.nickname = info.nickname
-            selfUser.value.grade = info.grade
-            getCardList()
-        }else{
-            Message.error("获取用户信息失败")
-        }
-    }catch(error){
-        console.error()
-    }
-})
-
-
-//获取打卡时长
-onMounted(async()=>{
-    const res = await checkoutInfo({token,id})
-    if(res.response?.code===200){
-        selfUser.value.totalDuration = res.response?.data.totalDuration
-        selfUser.value.targetDuration = res.response?.data.targetDuration
-        dataList.value.push(Math.floor(selfUser.value.totalDuration/60))
-    }
-})
-
-//获取打卡排名
-const getCardList = async()=>{
-    const response = await checkoutList({
-      grade: selfUser.value.grade as number,
-      pageSize: 40,
-      pageNum: 1
-    })
-    const responseData = response.response
-    console.log('responseData',responseData);
-    
-    if(responseData?.code===200){
-        userList.value = responseData.data.rows.map((row: any) => {
-        return {
-          avatar: row.avatar,
-          nickname: row.nickname,
-          totalDuration: (row.totalDuration / 60).toFixed(1),
-          targetDuration: row.targetDuration / 60,
-          grade: row.grade,
-          status: row.status
-        }
-      }) as User[]
-      userList.value.sort((a, b) => b.totalDuration - a.totalDuration)
-    }
-    userList.value.forEach((item,index)=>{
-        if(item.nickname===selfUser.value.nickname){
-            dataList.value.push(index+1)
-        }
-    })
-    dataList.value.push(0)
-}
+const { selfUser,dataList } = storeToRefs(userMain)
 
 </script>
